:recycle: Rename to components/atoms/buttons.dart
