+++ lib/initial_setting/initial_setting_2.dart
@@ -1,6 +1,6 @@
 import 'package:Pilll/application/router.dart';
 import 'package:Pilll/store/initial_setting.dart';
-import 'package:Pilll/style/button.dart';
+import 'package:Pilll/components/atoms/buttons.dart';
 import 'package:Pilll/components/atoms/color.dart';
 import 'package:Pilll/initial_setting/initial_setting_3.dart';
 import 'package:Pilll/components/organisms/pill/pill_sheet.dart';
